requirements.txt
biopython==1.70
certifi==2017.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
Flask==0.12.2
goatools==0.8.4
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
matplotlib==2.1.1
mygene==3.0.0
nose==1.3.7
numpy==1.13.3
pandas==0.21.0
patsy==0.4.1
py==1.4.34
pyparsing==2.2.0
pytest==3.2.3
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
six==1.11.0
sklearn==0.0
statsmodels==0.8.0
tornado==4.5.2
urllib3==1.22
Werkzeug==0.12.2
wget==3.2
xlrd==1.1.0
XlsxWriter==1.0.2
